SQL-Challenge Files
This folder includes images of the various codes/queries and the corresponding tables. There is also an image of the ERD sketches from QuickDBD, as well as the actual code used to create the queries and tables in pgAdmin4. There is also a README file explaining the contents and the assistance used during the assignment.